+++ flash_sparse_attn/ops/triton/flash_sparse_dec.py
@@ -618,9 +618,9 @@ def _flash_sparse_attn_decode(
     )
 
     if not is_quant:
-        query_scale = torch.ones(1, device=device, dtype=torch.float32)
-        key_scale = torch.ones(1, device=device, dtype=torch.float32)
-        value_scale = torch.ones(1, device=device, dtype=torch.float32)
+        query_scale = torch.ones(1, device=device, dtype=query.dtype)
+        key_scale = torch.ones(1, device=device, dtype=query.dtype)
+        value_scale = torch.ones(1, device=device, dtype=query.dtype)
 
     grid = launch_grid.get_dec_grid(
         batch_size=batch_size,
@@ -798,9 +798,9 @@ def _flash_sparse_attn_varlen_decode(
     )
 
     if not is_quant:
-        query_scale = torch.ones(1, device=device, dtype=torch.float32)
-        key_scale = torch.ones(1, device=device, dtype=torch.float32)
-        value_scale = torch.ones(1, device=device, dtype=torch.float32)
+        query_scale = torch.ones(1, device=device, dtype=query.dtype)
+        key_scale = torch.ones(1, device=device, dtype=query.dtype)
+        value_scale = torch.ones(1, device=device, dtype=query.dtype)
 
     grid = launch_grid.get_dec_grid(
         batch_size=batch_size,
